Picture's author and capture date is shown
git-svn-id: https://josm.openstreetmap.de/osmsvn/applications/editors/josm/plugins@31267 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ mapillary/src/org/openstreetmap/josm/plugins/mapillary/downloads/MapillaryImageInfoDownloaderThread.java
@@ -51,7 +51,7 @@ public class MapillaryImageInfoDownloaderThread implements Runnable {
 							((MapillaryImage) image).setUser(data
 									.getString("user"));
 							((MapillaryImage) image).setCapturedAt(data
-									.getJsonNumber("captured_at").intValue());
+									.getJsonNumber("captured_at").longValue());
 							((MapillaryImage) image).setLocation(data
 									.getString("location"));
 						}
